add tests

+++ Eventos/Eventos/Services/Enterprises/EnterpriseService.cs
@@ -9,6 +9,7 @@ namespace EventosApi.Services.Enterprises
     {
         private readonly IEnterpriseRepository _enterpriseRepository;
         private readonly IMapper _mapper;
+
         public EnterpriseService(IEnterpriseRepository EnterpriseRepository, IMapper mapper)
         {
             _enterpriseRepository = EnterpriseRepository;
@@ -20,27 +21,33 @@ namespace EventosApi.Services.Enterprises
             var enterprisesEntity = await _enterpriseRepository.GetAll();
             return _mapper.Map<IEnumerable<EnterpriseDto>>(enterprisesEntity);
         }
+
         public async Task<EnterpriseDto> GetEnterpriseById(Guid id)
         {
             var enterpriseEntity = await _enterpriseRepository.GetById(id);
             return _mapper.Map<EnterpriseDto>(enterpriseEntity);
         }
+
         public async Task AddEnterprise(EnterpriseDto enterpriseDto)
         {
             var enterpriseEntity = _mapper.Map<Enterprise>(enterpriseDto);
             await _enterpriseRepository.Create(enterpriseEntity);
             enterpriseDto.Id = enterpriseEntity.Id;
         }
+
         public async Task UpdateEnterprise(EnterpriseDto enterpriseDto)
         {
             var enterpriseEntity = _mapper.Map<Enterprise>(enterpriseDto);
             await _enterpriseRepository.Update(enterpriseEntity);
         }
-        public async Task RemoveEnterprise(Guid id)
+
+        public async Task<EnterpriseDto> RemoveEnterprise(Guid id)
         {
-            var enterpriseEntity = _enterpriseRepository.GetById(id).Result;
+            var enterpriseEntity = await _enterpriseRepository.GetById(id);
+            if (enterpriseEntity == null) return null;
+
             await _enterpriseRepository.Delete(enterpriseEntity.Id);
+            return _mapper.Map<EnterpriseDto>(enterpriseEntity);
         }
     }
 }
-
